Move ReReadingAdvisor to the chat advisor package
